kernel/fork: Boost CPU & DDR bus for a short amount of time when zygote forks
There is a gap from when a new app is forked to the time that the
devfreq governor managing the DDR bus can pick up the increased load and
ramp up appropriately. Since the polling interval for this governor is
typically 80 ms, that means it may take up to 80 ms before the governor
ramps up, which hurts cold launch app performance. Boosting the DDR bus
for 80 ms upon a fork from zygote covers that gap, speeding up cold app
launches.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: mcdofrenchfreis <xyzevan@androidist.net>